--- test/fixtures/array.prototype.findIndex/case-3/before.js
@@ -1,15 +1,0 @@
-var banana = require('array.prototype.findIndex');
-var assert = require('assert');
-
-assert.deepEqual(
-  banana([1, 2, 3], function (x) {
-    return x >= 2;
-  }),
-  [2, 3],
-);
-assert.deepEqual(
-  banana([1, 2, 3], function (x) {
-    return x <= 2;
-  }),
-  [1, 2],
-);
