+++ test/fixtures/array.prototype.findindex/case-1/before.js
@@ -1,0 +1,15 @@
+var findIndex = require('array.prototype.findindex');
+var assert = require('assert');
+
+assert.deepEqual(
+  findIndex([1, 2, 3], function (x) {
+    return x >= 2;
+  }),
+  [2, 3],
+);
+assert.deepEqual(
+  findIndex([1, 2, 3], function (x) {
+    return x <= 2;
+  }),
+  [1, 2],
+);
