+++ test/fixtures/array.prototype.findindex/case-1/result.js
