+++ test/fixtures/array.prototype.findindex/case-2/after.js
