+++ test/fixtures/array.prototype.findindex/case-3/after.js
