+++ test/fixtures/array.prototype.findindex/case-4/after.js
