+++ test/fixtures/array.prototype.findindex/case-4/result.js
