+++ test/fixtures/array.prototype.flatmap/case-1/after.js
