+++ test/fixtures/array.prototype.flatmap/case-1/result.js
