+++ test/fixtures/array.prototype.flatmap/case-2/after.js
