+++ test/fixtures/array.prototype.flatmap/case-2/before.js
@@ -1,0 +1,15 @@
+import flatMap from 'array.prototype.flatmap';
+var assert = require('assert');
+
+assert.deepEqual(
+  flatMap([1, 2, 3], function (x) {
+    return x >= 2;
+  }),
+  [2, 3],
+);
+assert.deepEqual(
+  flatMap([1, 2, 3], function (x) {
+    return x <= 2;
+  }),
+  [1, 2],
+);
