+++ test/fixtures/array.prototype.flatmap/case-2/result.js
