+++ test/fixtures/array.prototype.flatmap/case-3/after.js
