+++ test/fixtures/array.prototype.flatmap/case-3/before.js
@@ -1,0 +1,15 @@
+var banana = require('array.prototype.flatmap');
+var assert = require('assert');
+
+assert.deepEqual(
+  banana([1, 2, 3], function (x) {
+    return x >= 2;
+  }),
+  [2, 3],
+);
+assert.deepEqual(
+  banana([1, 2, 3], function (x) {
+    return x <= 2;
+  }),
+  [1, 2],
+);
