+++ test/fixtures/array.prototype.flatmap/case-3/result.js
