+++ test/fixtures/array.prototype.flatmap/case-4/after.js
