+++ test/fixtures/array.prototype.flatmap/case-4/result.js
