--- test/fixtures/array.prototype.forEach/case-2/before.js
@@ -1,15 +1,0 @@
-import forEach from 'array.prototype.forEach';
-var assert = require('assert');
-
-assert.deepEqual(
-  forEach([1, 2, 3], function (x) {
-    return x >= 2;
-  }),
-  [2, 3],
-);
-assert.deepEqual(
-  forEach([1, 2, 3], function (x) {
-    return x <= 2;
-  }),
-  [1, 2],
-);
