+++ test/fixtures/array.prototype.foreach/case-1/after.js
