+++ test/fixtures/array.prototype.foreach/case-1/result.js
