+++ test/fixtures/array.prototype.foreach/case-2/after.js
