+++ test/fixtures/array.prototype.foreach/case-2/result.js
