+++ test/fixtures/array.prototype.foreach/case-3/after.js
