+++ test/fixtures/array.prototype.foreach/case-3/result.js
