+++ test/fixtures/array.prototype.foreach/case-4/result.js
